Fixes requested by Jason:
The urls to the topics on the main search page are
not fully formed. It is breaking my link checking
software. The browsers are ok with it. It would be
good to fix this so that the automated testing tools
will work.

The section asking people to email you should be
called "We want to here from you!".  Ask people to
send us some feedback good or bad about codestriker.
We can get feedback this way, perhaps some nice,
quotes, and some company names.