Date: Mon, 28 Jun 2004 19:27:14 -0700 (PDT) From: Jason Remillard <jremillardshop@yahoo.com> To: David Sitsky <sits@users.sourceforge.net> Cc: Kannan Goundan <kannan@letterboxes.org>
Hi,

OK, it is broken if you have users hitting the topic that don't have
the email
cookie set. 1.8.1 would attempt to show the codestriker time as a
blank user if
it did not know who it was. 1.8.0 is broken for all multi user cases.

I tested the fix on a topic without an empty user name. This diff

1. Will fix that problem,
2. Show the empty user as (unknown user). Hopefully this will be
clearer.
3. Show the red feedback message at the top of the page when the user
saves the
metric data. I seem to remember you asking me to do this awhile
back. This
makes it consistent with the rest of CS.

Kannan you should be able to run patch against your latest from cvs to
get the
fix. It would be helpfull if you could apply it and see if it fixes
the problem
for you.

Sorry for the trouble, everybody...
Jason.